+++ packages/orq-rc/src/models/operations/updatetool.ts
@@ -1472,7 +1472,7 @@ export const RequestBodyTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   RequestBodyTools
 > = z.object({
-  id: z.string().default("01KRXPE59H38RX3EX8WF4TYKNT"),
+  id: z.string().default("01KRZ9VJYQ6XZ3V6CPSA4G5WTE"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => UpdateToolRequestBodyToolsSchema$outboundSchema),
@@ -2193,7 +2193,7 @@ export const UpdateToolResponseBodyCodeExecutionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KRXPE59DYFNR7AACX74ERCK6"),
+  _id: z.string().default("tool_01KRZ9VJYKM10XZTZW13RY0DS1"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2296,7 +2296,7 @@ export const UpdateToolResponseBodyTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KRXPE59CE7S0J5BSGNQDYCM7"),
+  id: z.string().default("01KRZ9VJYJMGEC3Z1TV3CWJXJ6"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => UpdateToolResponseBodyToolsSchema$inboundSchema),
@@ -2354,7 +2354,7 @@ export const UpdateToolResponseBodyMCPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KRXPE59B1HEFMZYG21Y5H5QP"),
+  _id: z.string().default("tool_01KRZ9VJYHQ5XN6DSVJX5E9ZT4"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2543,7 +2543,7 @@ export const UpdateToolResponseBodyHTTPTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KRXPE59AFBEANJD6FT33BG9E"),
+  _id: z.string().default("tool_01KRZ9VJYFGCEY8MXHC50JDBXP"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2638,7 +2638,7 @@ export const UpdateToolResponseBodyJSONSchemaTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KRXPE5993DDVAT5TSAG19J9F"),
+  _id: z.string().default("tool_01KRZ9VJYEJX6H7SBXZ9T948KC"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),
@@ -2741,7 +2741,7 @@ export const UpdateToolResponseBodyFunctionTool$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("tool_01KRXPE598NC2BH43E58KASD93"),
+  _id: z.string().default("tool_01KRZ9VJYDFECCTRHG52GGQ4P2"),
   path: z.string(),
   key: z.string(),
   display_name: z.string().optional(),